consumer/requirements.txt
hdfs==2.7.3
kafka-python==2.1.1